JavaScript Arrow Function
Arrow Functions:
The msg function is an arrow function that logs a message, Hello JS, to the console.
Arrow functions provide a more concise syntax for writing functions.

Return Type Functions:
The add function is an arrow function that returns the sum of two numbers. It illustrates how functions can return values.

Default Parameters:
The mult function shows how you can set default values for parameters (e.g., b = 2). If no second argument is provided, it defaults to 2.

Rest Parameters:
The mysum function uses the rest parameter (...n) to accept a variable number of arguments, which are then printed as an array.

Using reduce for Summing Values:
The total function calculates the sum of all arguments passed to it using the reduce method. This is a concise and functional approach to accumulating values from an array (or rest parameters in this case).